src/requirements.txt
astroid==1.6.4
colorama==0.3.9
coverage==4.5.1
Django==2.0.5
django-cors-headers==2.2.0
django-nose==1.4.5
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysql-connector-python==8.0.11
mysqlclient==1.3.12
nose==1.3.7
pinocchio==0.4.2
protobuf==3.5.2.post1
pycodestyle==2.4.0
pylint==1.9.1
pylint-django==0.11.1
pylint-plugin-utils==0.2.6
pytz==2018.4
six==1.11.0
wrapt==1.10.11
